rename module: models -> schema
